+++ src/components/index.ts
@@ -1,3 +1,1 @@
-export * from './Button.component';
-export * from './Button.types';
-export * from './Button.constants';
+export * from './Buttons';
