+++ tsconfig.json
@@ -18,7 +18,7 @@
     "dist",
     "node_modules",
     "src/**/*.test.tsx",
-//    "src/**/*.test.ts",
+    "src/**/*.test.ts",
     "src/**/*.stories.tsx",
   ],
 }
